requirements.txt
blinker==1.7.0
boto3==1.34.59
botocore==1.34.59
click==8.1.7
colorama==0.4.6
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.0
six==1.16.0
urllib3==2.0.7
Werkzeug==3.0.1
